# perf(clustering): no flattening on data plane 