--- a/kong-3.5.0-0.rockspec
+++ b/kong-3.5.0-0.rockspec
@@ -100,6 +100,8 @@ build = {
     ["kong.resty.mlcache"] = "kong/resty/mlcache/init.lua",
     ["kong.resty.mlcache.ipc"] = "kong/resty/mlcache/ipc.lua",
 
+    ["kong.resty.lmdb.reset-transaction"] = "kong/resty/lmdb/reset-transaction.lua",
+
     ["kong.cmd"] = "kong/cmd/init.lua",
     ["kong.cmd.roar"] = "kong/cmd/roar.lua",
     ["kong.cmd.stop"] = "kong/cmd/stop.lua",