--- a/kong/db/declarative/init.lua
+++ b/kong/db/declarative/init.lua
@@ -244,6 +244,7 @@ end
 _M.to_yaml_string              = declarative_export.to_yaml_string
 _M.to_yaml_file                = declarative_export.to_yaml_file
 _M.export_from_db              = declarative_export.export_from_db
+_M.export                      = declarative_export.export
 _M.export_config               = declarative_export.export_config
 _M.export_config_proto         = declarative_export.export_config_proto
 _M.sanitize_output             = declarative_export.sanitize_output