### Summary

A quick unscientific test shows:

`300.000` routes and services, and adding or deleting a route (around `600.000` entities in total).

#### master:
```
1st sync: declarative import took 27819 ms
2nd sync: declarative import took 32424 ms
```

#### perf/no-flattening:
```
1st sync: declarative import took 18441 ms
2nd sync: declarative import took 21658 ms
```

So it reduces the CPU time in this massive data set by 10 secs, or close to 1/3. Non-yieldable LMDB commit is 2-3 secs on `perf/no-flattening`.

KAG-2676