@bungle 2023-09-28
If we can remove the `validate_entity` we get down to `5399 ms` for import.